# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/maintenance/chart-of-accounts": {
      "get": {
        "operationId": "ChartOfAccountsController_findAll_v1",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 120,
              "type": "string"
            }
          },
          {
            "name": "accountLevel",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "SUB3",
                "SUB1",
                "SUB2",
                "MAJOR",
                "SPECIFIC"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "INACTIVE"
              ]
            }
          },
          {
            "name": "parentAccountId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChartAccountListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Chart of Accounts"
        ]
      },
      "post": {
        "operationId": "ChartOfAccountsController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChartAccountDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChartAccountSaveResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Chart of Accounts"
        ]
      }
    },
...
    "/api/v1/maintenance/chart-of-accounts/next-code": {
      "get": {
        "operationId": "ChartOfAccountsController_findNextCode_v1",
        "parameters": [
          {
            "name": "parentAccountId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accountLevel",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "SUB3",
                "SUB1",
                "SUB2",
                "MAJOR",
                "SPECIFIC"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChartAccountNextCodeResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Chart of Accounts"
        ]
      }
    },
    "/api/v1/maintenance/chart-of-accounts/{id}": {
      "get": {
        "operationId": "ChartOfAccountsController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChartAccountContainerResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Chart of Accounts"
        ]
      },
      "patch": {
        "operationId": "ChartOfAccountsController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateChartAccountDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChartAccountSaveResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Chart of Accounts"
        ]
      }
    },
    "/api/v1/maintenance/chart-of-accounts/{id}/status": {
      "patch": {
        "operationId": "ChartOfAccountsController_updateStatus_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateChartAccountStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChartAccountSaveResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Chart of Accounts"
        ]
      }
    },
...
    "/api/v1/maintenance/discount-maintenance": {
      "get": {
        "operationId": "DiscountMaintenanceController_findAll_v1",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 120,
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "SALES",
                "PURCHASE"
              ]
            }
          },
          {
            "name": "valueType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PERCENTAGE",
                "FIXED"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "INACTIVE"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 500,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "name",
                "type",
                "valueType",
                "value",
                "status",
                "createdAt",
                "updatedAt"
              ]
            }
          },
          {
            "name": "sortDirection",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Discount list retrieved."
          }
        },
        "tags": [
          "Discount Maintenance"
        ]
      },
      "post": {
        "operationId": "DiscountMaintenanceController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDiscountDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Discount created."
          }
        },
        "tags": [
          "Discount Maintenance"
        ]
      }
    },
    "/api/v1/maintenance/discount-maintenance/{id}": {
      "get": {
        "operationId": "DiscountMaintenanceController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Discount retrieved."
          }
        },
        "tags": [
          "Discount Maintenance"
        ]
      },
      "patch": {
        "operationId": "DiscountMaintenanceController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDiscountDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Discount updated."
          }
        },
        "tags": [
          "Discount Maintenance"
        ]
      }
    },
    "/api/v1/maintenance/discount-maintenance/import": {
      "post": {
        "operationId": "DiscountMaintenanceController_importDiscounts_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportDiscountsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Discounts imported."
          }
        },
        "tags": [
          "Discount Maintenance"
        ]
      }
    },
...
    "/api/v1/maintenance/financial-management/default-accounts": {
      "get": {
        "operationId": "DefaultAccountController_findAll_v1",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 120,
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "EXPENSE",
                "COLLECTION",
                "FIXED_ASSET"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "INACTIVE"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 500,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "name",
                "description",
                "type",
                "status",
                "createdAt",
                "updatedAt"
              ]
            }
          },
          {
            "name": "sortDirection",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Default accounts retrieved."
          }
        },
        "tags": [
          "Default Account"
        ]
      },
      "post": {
        "operationId": "DefaultAccountController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDefaultAccountTemplateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default account created."
          }
        },
        "tags": [
          "Default Account"
        ]
      }
    },
...
    "/api/v1/maintenance/financial-management/default-accounts/{id}": {
      "get": {
        "operationId": "DefaultAccountController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Default account retrieved."
          }
        },
        "tags": [
          "Default Account"
        ]
      },
      "patch": {
        "operationId": "DefaultAccountController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDefaultAccountTemplateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default account updated."
          }
        },
        "tags": [
          "Default Account"
        ]
      }
    },
    "/api/v1/maintenance/financial-management/default-accounts/{id}/status": {
      "patch": {
        "operationId": "DefaultAccountController_updateStatus_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDefaultAccountTemplateStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default account status updated."
          }
        },
        "tags": [
          "Default Account"
        ]
      }
    },
...
  },
  "info": {
    "title": "GR8Books Neo API",
    "description": "Versioned REST API for GR8Books Neo frontend clients.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ChartAccountListResponseDto": {
        "type": "object",
        "properties": {
          "accounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChartAccountResponseDto"
            }
          }
        },
        "required": [
          "accounts"
        ]
      },
...
      "ChartAccountNextCodeResponseDto": {
        "type": "object",
        "properties": {
          "accountCode": {
            "type": "string"
          }
        },
        "required": [
          "accountCode"
        ]
      },
      "ChartAccountContainerResponseDto": {
        "type": "object",
        "properties": {
          "account": {
            "$ref": "#/components/schemas/ChartAccountResponseDto"
          }
        },
        "required": [
          "account"
        ]
      },
      "ChartAccountBankDetailsDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "BANK"
            ]
          },
          "bankName": {
            "type": "string",
            "maxLength": 100
          },
          "branch": {
            "type": "string",
            "maxLength": 100
          },
          "accountNumber": {
            "type": "string",
            "maxLength": 100
          },
          "accountType": {
            "type": "string",
            "maxLength": 50
          },
          "currencyCode": {
            "type": "string",
            "maxLength": 10
          },
          "currencyExchangeRate": {
            "type": "number",
            "minimum": 0
          },
          "seriesStart": {
            "type": "string",
            "maxLength": 50
          },
          "seriesEnd": {
            "type": "string",
            "maxLength": 50
          },
          "seriesDigits": {
            "type": "number",
            "minimum": 1
          }
        }
      },
      "CreateChartAccountDto": {
        "type": "object",
        "properties": {
          "parentAccountId": {
            "type": "string"
          },
          "accountLevel": {
            "enum": [
              "SUB3",
              "SUB1",
              "SUB2",
              "MAJOR",
              "SPECIFIC"
            ],
            "type": "string"
          },
          "accountTitle": {
            "type": "string",
            "maxLength": 250
          },
          "accountType": {
            "enum": [
              "ASSET",
              "EXPENSE",
              "REVENUE",
              "LIABILITY",
              "EQUITY"
            ],
            "type": "string"
          },
          "accountNature": {
            "enum": [
              "DEBIT",
              "CREDIT"
            ],
            "type": "string"
          },
          "accountGroup": {
            "type": "object"
          },
          "statementSection": {
            "type": "string",
            "maxLength": 250
          },
          "reportAlias": {
            "type": "string",
            "maxLength": 250
          },
          "description": {
            "type": "string",
            "maxLength": 500
          },
          "isPostingAccount": {
            "type": "boolean"
          },
          "withSubsidiary": {
            "type": "boolean"
          },
          "contraAccount": {
            "type": "boolean"
          },
          "showTotal": {
            "type": "boolean"
          },
          "orderNo": {
            "type": "number",
            "minimum": 0
          },
          "currencyCode": {
            "type": "string",
            "maxLength": 10
          },
          "status": {
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ],
            "type": "string"
          },
          "linkedDetails": {
            "$ref": "#/components/schemas/ChartAccountBankDetailsDto"
          }
        },
        "required": [
          "accountLevel",
          "accountTitle"
        ]
      },
      "ChartAccountSaveResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "account": {
            "$ref": "#/components/schemas/ChartAccountResponseDto"
          }
        },
        "required": [
          "message",
          "account"
        ]
      },
      "UpdateChartAccountDto": {
        "type": "object",
        "properties": {
          "parentAccountId": {
            "type": "string"
          },
          "accountLevel": {
            "enum": [
              "SUB3",
              "SUB1",
              "SUB2",
              "MAJOR",
              "SPECIFIC"
            ],
            "type": "string"
          },
          "accountTitle": {
            "type": "string",
            "maxLength": 250
          },
          "accountType": {
            "enum": [
              "ASSET",
              "EXPENSE",
              "REVENUE",
              "LIABILITY",
              "EQUITY"
            ],
            "type": "string"
          },
          "accountNature": {
            "enum": [
              "DEBIT",
              "CREDIT"
            ],
            "type": "string"
          },
          "accountGroup": {
            "type": "object"
          },
          "statementSection": {
            "type": "string",
            "maxLength": 250
          },
          "reportAlias": {
            "type": "string",
            "maxLength": 250
          },
          "description": {
            "type": "string",
            "maxLength": 500
          },
          "isPostingAccount": {
            "type": "boolean"
          },
          "withSubsidiary": {
            "type": "boolean"
          },
          "contraAccount": {
            "type": "boolean"
          },
          "showTotal": {
            "type": "boolean"
          },
          "orderNo": {
            "type": "number",
            "minimum": 0
          },
          "currencyCode": {
            "type": "string",
            "maxLength": 10
          },
          "status": {
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ],
            "type": "string"
          },
          "linkedDetails": {
            "$ref": "#/components/schemas/ChartAccountBankDetailsDto"
          }
        }
      },
      "UpdateChartAccountStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ],
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
...
      "CreateDiscountDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 150
          },
          "description": {
            "type": "string",
            "maxLength": 500
          },
          "type": {
            "enum": [
              "SALES",
              "PURCHASE"
            ],
            "type": "string"
          },
          "valueType": {
            "enum": [
              "PERCENTAGE",
              "FIXED"
            ],
            "type": "string"
          },
          "value": {
            "type": "number",
            "minimum": 0
          },
          "status": {
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ],
            "type": "string"
          }
        },
        "required": [
          "name",
          "type",
          "valueType",
          "value"
        ]
      },
      "ImportDiscountsDto": {
        "type": "object",
        "properties": {
          "discounts": {
            "minItems": 1,
            "maxItems": 500,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateDiscountDto"
            }
          }
        },
        "required": [
          "discounts"
        ]
      },
      "UpdateDiscountDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 150
          },
          "description": {
            "type": "string",
            "maxLength": 500
          },
          "type": {
            "enum": [
              "SALES",
              "PURCHASE"
            ],
            "type": "string"
          },
          "valueType": {
            "enum": [
              "PERCENTAGE",
              "FIXED"
            ],
            "type": "string"
          },
          "value": {
            "type": "number",
            "minimum": 0
          },
          "status": {
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ],
            "type": "string"
          }
        }
      },
...
      "CreateDefaultAccountTemplateDto": {
        "type": "object",
        "properties": {
          "type": {
            "enum": [
              "EXPENSE",
              "COLLECTION",
              "FIXED_ASSET"
            ],
            "type": "string"
          },
          "defaultAccountName": {
            "type": "string",
            "maxLength": 250
          },
          "description": {
            "type": "string",
            "maxLength": 500
          },
          "status": {
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ],
            "type": "string"
          },
          "expenseParentCoaId": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "defaultAccountName"
        ]
      },
      "UpdateDefaultAccountTemplateDto": {
        "type": "object",
        "properties": {
          "type": {
            "enum": [
              "EXPENSE",
              "COLLECTION",
              "FIXED_ASSET"
            ],
            "type": "string"
          },
          "defaultAccountName": {
            "type": "string",
            "maxLength": 250
          },
          "description": {
            "type": "string",
            "maxLength": 500
          },
          "status": {
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ],
            "type": "string"
          },
          "expenseParentCoaId": {
            "type": "string"
          }
        }
      },
      "UpdateDefaultAccountTemplateStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ],
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}